requirements.txt
sqlmodel
typing-extensions
jiwer
gradio
fastapi
sqlalchemy
loguru
pytest
httpx
python-dotenv
psycopg2
